+++ src/main/java/com/plivo/api/models/number/PhoneNumberCreationStatus.java
@@ -4,6 +4,7 @@ public class PhoneNumberCreationStatus {
 
   private String number;
   private String status;
+  private String fallbackNumber;
 
   public String getNumber() {
     return number;
@@ -11,5 +12,9 @@ public class PhoneNumberCreationStatus {
 
   public String getStatus() {
     return status;
+  }
+
+  public String getFallbackNumber() {
+    return fallbackNumber;
   }
 }