+++ src/main/java/com/plivo/api/models/number/PhoneNumberCreator.java
@@ -8,6 +8,7 @@ public class PhoneNumberCreator extends Creator<PhoneNumberCreateResponse> {
   private String number;
   private String appId;
   private String cnamLookup;
+  private Boolean haEnable;
 
   public PhoneNumberCreator(String number) {
     if (number == null) {
@@ -32,6 +33,15 @@ public class PhoneNumberCreator extends Creator<PhoneNumberCreateResponse> {
 
   public PhoneNumberCreator cnamLookup(final String cnamLookup) {
     this.cnamLookup = cnamLookup;
+    return this;
+  }
+
+  public Boolean haEnable() {
+    return this.haEnable;
+  }
+
+  public PhoneNumberCreator haEnable(final Boolean haEnable) {
+    this.haEnable = haEnable;
     return this;
   }